Fix for double Invoice Card/Bankpay new Api
When callback is to fast. Lets say 0.1s the AcceptUrl isnt finished and callback also adds a invoice because of order statsus is still not changed.

The fix solves this with Wordpress Transients Api to set a lockstate on order when for example accept is processing prevents other actions in the same time.